Type 0,1,2 queries implemented

+++ linear-road-flink/src/main/java/com/github/wladox/model/XwayDirSeg.java
@@ -2,9 +2,9 @@ package com.github.wladox.model;
 
 import org.apache.flink.api.java.tuple.Tuple3;
 
-public class XwayDirSeg extends Tuple3<Byte, Byte, Byte> {
+public class XwayDirSeg extends Tuple3<Integer, Integer, Integer> {
 
-  public XwayDirSeg(Byte xWay, Byte direction, Byte segment) {
+  public XwayDirSeg(Integer xWay, Integer direction, Integer segment) {
     super(xWay, direction, segment);
   }
 }
